Add files via upload
fixed floating point comparison bug

+++ PrimRootFinder.java
@@ -85,7 +85,7 @@ public class PrimRootFinder {
 
 
 
-    public boolean isPrime(double x){
+    public boolean isPrime(int x){
         if(x<2){return false;}
         for (int i = 2; i <= x/2; i++) {
             if(x%i == 0){ return false; }
